Added new utility class `Views::Things::Shared::MessageEncoder`.
When you assign a flash message, say, `flash[:error] = 'Oops.'`, it
overwrites anything that had previously been assigned to
`flash[:error]`, just as you’d expect a Hash-like thing to work. Fine;
that means that if you have multiple errors, you need to encode them
using something like JSON. Not a big deal, as far as that goes. We
have two flash-related classes, `Views::Things::Shared::Flashes` and
`Views::Things::Shared::FlashAlert; the former simply iterates through
the passed-in flash-message Hash-like thing and creates one instance
of the latter for each level/message pair, which renders an
appropriately-styled alert box for each message. ​_These classes assume
that the “message”-like thing is in fact a single message string, per
what I said at the beginning._​ If we in fact want to generate multiple
alerts for multiple messages, then the responder needs to assign a
serialised (e.g., JSON-encoded) string representing the collection of
messages. The logical place to decode that list would be a new class,
between the existing `Flashes` and `FlashAlert`. The ​_safe_​ thing to
do would be for that new layer to attempt to decode the incoming
string. If decoding fails, then the passed-in message string is just a
string, and we pass it on as now. Otherwise, we take the decoded
message (which might be a String or an Array of Strings), and coerce
it to an Array, then iterate that array. That would give a reasonably
robust implementation of a new inner layer that allows us to sandwich
it between the two existing layers, modified only in that the outer
layer calls the new middle layer instead of the inner layer.

This new `Views::Things::Shared::MessageEncoder` class provides that
"middle layer"; code setting a flash message (e.g., in a responder)
would call its `.encode` method and be handed back an encoded string
(JSON-encoded, but nobody on either side of the sandwich cares) and
assign it to, e.g., `flash[:error]`; view code would take the string
that had been assigned to a flash message and call `.decode`, knowing
that it will get back an array of zero or more strings. It would then
iterate that array using the existing `FlashAlert` class. This would
require a change to the existing `Flashes` class (that presently does
that iteration), which should be noted in any after-action review as
an undesirable effect (but manageable in this case).

159 tests, 220 assertions, 0 failures, 0 errors, 0 skips
Coverage: 382 / 382 LOC (100.0%) covered
RuboCop: 48 files inspected, no offenses detected
Flay: Total score 42
Flog: Total 309.6; method average 3.2; max 7.0 (Views::Things::New#flash_messages)
Reek: 0 warnings